Enhanced Music Player with Dynamic Playlist and Scroll Animation
This implements dynamic playlist fetching and scroll animation features for a music player. The playlist songs are fetched from a server and added to the playlist dynamically. If a song is not available in the playlist, it will be fetched and added on-the-fly when played.

To enhance the user experience, a scroll animation is added to the playlist only if the song subtitle width exceeds the song title width. This prevents overlapping and ensures better visibility of the song information.

Additionally, the project addresses several other improvements.
1. To avoid redirection to the Spotify app, meta headers are hidden.
2. To eliminate 404 errors, dummy images and song titles are removed.
3. To prevent subtitle overlap, the song title tag is set to have overflow scroll on the x-axis with a fixed width.
4. And also resolves the issue of adding the playlist to a queue by directly playing the first song.